+++ src/components/content/conditions/patient-conditions-filters.tsx
@@ -1,20 +1,66 @@
 import { useState } from "react";
+import {
+  FilterChangeEvent,
+  FilterItem,
+} from "@/components/core/filter-bar/filter-bar-types";
 import { ConditionModel } from "@/fhir/models";
-import { cloneDeep, merge } from "@/utils/nodash";
+import { compact, isArray, uniq } from "@/utils/nodash";
 
 export type FilterCollection = "patient" | "other";
 
 export type Filters = {
-  showHistoric: boolean;
+  patient: FilterChangeEvent;
+  other: FilterChangeEvent;
+};
+
+const DEFAULT_FILTERS: Filters = {
+  patient: {
+    displayStatus: {
+      key: "displayStatus",
+      selected: ["Active", "Pending", "Unknown"],
+      type: "checkbox",
+    },
+  },
+  other: {
+    displayStatus: {
+      key: "displayStatus",
+      selected: ["Active", "Pending", "Unknown"],
+      type: "checkbox",
+    },
+  },
 };
 
 export function useConditionFilters(collection: FilterCollection) {
   const [filters, setFilters] = useState<Filters>({
-    showHistoric: false,
+    ...DEFAULT_FILTERS,
   });
 
   function updateFilters(newFilters: Partial<Filters>) {
-    setFilters(merge(cloneDeep(filters), newFilters));
+    setFilters(() => ({
+      ...filters,
+      [collection]: {
+        ...newFilters,
+      },
+    }));
+  }
+
+  function availableFilters(conditions: ConditionModel[]): FilterItem[] {
+    return [
+      {
+        key: "displayStatus",
+        type: "checkbox",
+        icon: "eye",
+        display: () => "Status",
+        values: uniq(conditions.map((c) => c.displayStatus)),
+      },
+      {
+        key: "ccsChapter",
+        type: "checkbox",
+        icon: "eye",
+        display: () => "Category",
+        values: compact(uniq(conditions.map((c) => c.ccsChapter))),
+      },
+    ];
   }
 
   function applyFilters(
@@ -23,12 +69,36 @@ export function useConditionFilters(collection: FilterCollection) {
   ) {
     const conditions =
       collection === "patient" ? patientConditions : otherConditions;
-    return conditions.filter((c) => {
-      if (filters.showHistoric) return true;
 
-      return ["Active", "Pending"].includes(c.displayStatus);
-    });
+    return conditions.filter((condition) =>
+      Object.entries(filters[collection]).every(([_, filterItem]) => {
+        if (filterItem?.type === "checkbox" && isArray(filterItem.selected)) {
+          const filteredList = filterItem.selected.filter((item) =>
+            compact(
+              uniq(
+                conditions.map((c) => c[filterItem.key as keyof ConditionModel])
+              )
+            ).includes(item)
+          );
+
+          const targetFilter =
+            condition[filterItem.key as keyof ConditionModel] ?? "";
+
+          return (
+            filteredList.length < 1 ||
+            filteredList.includes(String(targetFilter))
+          );
+        }
+
+        return true;
+      })
+    );
   }
 
-  return { filters, updateFilters, applyFilters };
+  return {
+    filters,
+    updateFilters,
+    applyFilters,
+    availableFilters,
+  };
 }