+++ src/components/content/conditions/patient-conditions.tsx
@@ -15,33 +15,31 @@ import {
 import { useConditionSorts } from "./patient-conditions-sort";
 import { PatientConditionsTabs } from "./patient-conditions-tabs";
 import { withErrorBoundary } from "@/components/core/error-boundary";
-import { FormEntry } from "@/components/core/form/drawer-form-with-fields";
 import { Table } from "@/components/core/table/table";
 import {
   useOtherProviderConditions,
   usePatientConditions,
 } from "@/fhir/conditions";
 import { useBreakpoints } from "@/hooks/use-breakpoints";
-import { AnyZodSchema } from "@/utils/form-helper";
 import "./patient-conditions.scss";
 
 export type PatientConditionsProps = {
   className?: string;
   readOnly?: boolean;
-};
-
-export type ConditionFormData = {
-  schema: AnyZodSchema;
-  actionType: string;
-  data: FormEntry[] | undefined;
-  drawerIsOpen: boolean;
+  hideBuilderOwnedRecords?: boolean;
 };
 
 export const PatientConditions = withErrorBoundary(
-  ({ className, readOnly = false }: PatientConditionsProps) => {
+  ({
+    className,
+    readOnly = false,
+    hideBuilderOwnedRecords = false,
+  }: PatientConditionsProps) => {
     // State.
-    const [collection, setCollection] = useState<FilterCollection>("patient");
-    const { filters, updateFilters, applyFilters } =
+    const [collection, setCollection] = useState<FilterCollection>(
+      hideBuilderOwnedRecords ? "other" : "patient"
+    );
+    const { filters, updateFilters, applyFilters, availableFilters } =
       useConditionFilters(collection);
     const { applySorts, sortOptions, updateSorts, currentSorts } =
       useConditionSorts(collection);
@@ -68,6 +66,7 @@ export const PatientConditions = withErrorBoundary(
       patientConditions,
       true
     );
+
     let conditions = applyFilters(patientConditions, otherConditions);
     conditions = applySorts(conditions);
     const RowActions =
@@ -78,48 +77,53 @@ export const PatientConditions = withErrorBoundary(
     return (
       <div
         ref={containerRef}
-        className={cx("ctw-patient-conditions", className, {
-          "ctw-patient-conditions-stacked": breakpoints.sm,
-        })}
+        className={cx(
+          "ctw-patient-resource-component ctw-patient-conditions ctw-items-center ctw-justify-between ctw-py-5",
+          className,
+          {
+            "ctw-patient-conditions-stacked": breakpoints.sm,
+          }
+        )}
       >
-        <div className="ctw-items-center ctw-justify-between ctw-py-5">
-          <div className="ctw-ml-3 ctw-text-xl ctw-font-medium ctw-text-content-black">
-            Conditions
-          </div>
+        {!hideBuilderOwnedRecords && (
           <PatientConditionsTabs
+            forceHorizontalTabs
             otherConditions={otherConditions}
             collection={collection}
-            onCollectionChange={(c) => setCollection(c)}
+            onCollectionChange={setCollection}
           />
+        )}
 
-          <PatientConditionsActions
-            sortOptions={sortOptions}
-            updateSorts={updateSorts}
-            activeCollection={collection}
-            hideAdd={readOnly || collection === "other"}
-            onToggleShowHistoric={() =>
-              updateFilters({ showHistoric: !filters.showHistoric })
-            }
-            currentSorts={currentSorts[collection]}
-          />
+        <PatientConditionsActions
+          sortOptions={sortOptions}
+          updateSorts={updateSorts}
+          activeCollection={collection}
+          hideAdd={readOnly || collection === "other"}
+          currentSorts={currentSorts[collection]}
+          filterItems={availableFilters(
+            collection === "patient" ? patientConditions : otherConditions
+          )}
+          setFilters={updateFilters}
+          filters={filters[collection]}
+        />
 
-          <Table
-            stacked={breakpoints.sm}
-            className="-ctw-mx-px !ctw-rounded-none"
-            showTableHead={false}
-            emptyMessage="There are no condition records available."
-            isLoading={isLoading()}
-            records={conditions}
-            RowActions={readOnly ? undefined : RowActions}
-            columns={patientConditionsColumns}
-            handleRowClick={(condition) =>
-              showConditionHistory({
-                condition,
-                readOnly: readOnly || condition.isSummaryResource,
-              })
-            }
-          />
-        </div>
+        <Table
+          stacked={breakpoints.sm}
+          removeLeftAndRightBorders
+          className="-ctw-mx-px !ctw-rounded-none"
+          showTableHead={false}
+          emptyMessage="There are no condition records available."
+          isLoading={isLoading()}
+          records={conditions}
+          RowActions={readOnly ? undefined : RowActions}
+          columns={patientConditionsColumns}
+          handleRowClick={(condition) =>
+            showConditionHistory({
+              condition,
+              readOnly: readOnly || condition.isSummaryResource,
+            })
+          }
+        />
       </div>
     );
   },